plugins: Load and register plugins separately
When attaching a plugin to Ossaca, the plugin's default config might not
allow the plugin to work properly. We must however give a chance for the
plugin to be configured and reloaded. To do so, allow to re-load a
plugin after it has been attached.

This can be known by looking at the plugin state.

Signed-off-by: Maxime Chevallier <chevallier.seb.max@gmail.com>